requirements.txt
customtkinter
pygame
pillow
packaging